refactor(tests): co-locate all tests under src/__tests__/
- Move analytics tests (google, umami, AnalyticsScripts) to src/__tests__/analytics/
- Move component tests (Avatar, Button, Home, Share, Sort) to src/__tests__/components/
- Move runtimeConfig.test.js to src/__tests__/config/
- Update all relative import paths in moved test files
- Add .prettierrc config matching eslint prettier rule (singleQuote: true)
- Fix public/css/brands.css: change button display from inline-block to block
  to prevent side-by-side layout on wide screens
