+++ .prettierrc
@@ -1,0 +1,7 @@
+{
+  "singleQuote": true,
+  "jsxSingleQuote": false,
+  "arrowParens": "avoid",
+  "semi": true,
+  "trailingComma": "all"
+}
