+++ src/__tests__/components/Button/Button.test.js
@@ -1,4 +1,4 @@
-import Button from '../Button';
+import Button from '../../../components/Button/Button';
 import React from 'react';
 import { render } from '@testing-library/react';
